Refactor lang item mappings enum into its own header
